BE/requirements.txt
flask
mysql-connector-python
bcrypt
flask-cors
python-dotenv